TrustInterfaceService: Do the needful at a later stage
According to the lifecycle of a SystemService, onStart() is called to get
the service running and the service should publish its binder interface
at that point. First of all, ensure to do that and then do the real work
once boot is completed.

This was actually the root cause of the issue that commit c06e32d
addressed (not entirely correctly). Although the crash was no longer
observed after that commit, the proper fix is actually this one.

Change-Id: I42dc7182ba3d0db5413709de229ab9c0c4926df6